Comment

+++ routers/postRouter.js
@@ -19,9 +19,9 @@ postRouter.get('/',async(req,res)=>{
     const datas = await getUserSpecificData(req)
     var posts = null
     if(datas.user){
-        posts = await Post.find({$or:[{posted:true,showTo:{"$in":[datas.user._id]}},{posted:true,userId:datas.user._id}]}).sort('-updatedAt')
+        posts = await Post.find({$or:[{posted:true,showTo:{"$in":[datas.user._id]}},{posted:true,userId:datas.user._id}]}).sort('-updatedAt').select('_id user title description youtubeId files likes createdAt commentCount userId')
     }else{
-        posts = await Post.find({posted:true})
+        posts = await Post.find({posted:true}).select('_id user title description youtubeId files likes createdAt commentCount userId')
     }
     res.send(posts)
 })
@@ -67,21 +67,62 @@ postRouter.get('/like/:p_id',isAuth,async(req,res)=>{
 postRouter.post('/update/:p_id',isAuth, async(req,res)=>{
     const post_id = req.params['p_id']
     const post = await Post.findById(post_id)
-    // if(post.userId != req.user._id) res.send({"Error":"You are not the owner"})
-    const data = req.body
-    post.title = data.title
-    post.description = data.description
-    post.youtubeId = data.youtubeId
-    if(post.title && (post.youtubeId || post.description)){
-        post.posted = true
-    }else{
-        await Post.deleteMany({_id:{$ne:post_id},userId:req.user._id,posted:false})
-        post.posted = false
+    if(post.userId === req.user._id){
+        const data = req.body
+        post.title = data.title
+        post.description = data.description
+        post.youtubeId = data.youtubeId
+        if(post.title && (post.youtubeId || post.description)){
+            post.posted = true
+        }else{
+            await Post.deleteMany({_id:{$ne:post_id},userId:req.user._id,posted:false})
+            post.posted = false
+        }
+        await post.save()
     }
-    await post.save()
     res.send({"post":post})
 })
 
+postRouter.get('/comments/:p_id',async(req,res)=>{
+    const p_id = req.params['p_id']
+    var post = await Post.findById(p_id)
+    if(!post.comments){
+        post.comments = []
+    }
+    res.send(post.comments)
+})
 
+postRouter.post('/add_comment/:p_id',isAuth,async(req,res) => {
+    const p_id = req.params['p_id']
+    var post = await Post.findById(p_id)
+    const user = await User.findById(req.user._id)
+    const data = req.body
+    var comment = {
+        userId:user.userId,
+        name:user.name,
+        avatar:user.avatar,
+        content:data.content
+    }
+    if(!post.comments){
+        post.comments = []
+    }
+    post.comments=[...post.comments,comment]
+    post.commentCount = post.comments.length
+    await post.save()
+    res.send(post.comments)
+})
+
+postRouter.get('/delete_comment/:p_id/:c_id',isAuth,async(req,res)=>{
+    const p_id = req.params['p_id']
+    const c_id = req.params['c_id']
+    var post = await Post.findById(p_id)
+    const index = post.comments.findIndex(comment=>comment._id.toString() === c_id)
+    if (index>-1){
+        post.comments.splice(index,1)
+    }
+    post.commentCount = post.comments.length
+    await post.save()
+    res.send(post.comments)
+})
 
 module.exports = postRouter
